.: Split out unit tests into two files (Chapter 12)
