+++ changeLog.txt
@@ -8,6 +8,7 @@ develop
 -------------------------------------------------------------
 .: Moved Fts into their own individual files (Chapter 12)
 .: Move unit tests into a folder with single file (Chapter 12)
+.: Split out unit tests into two files (Chapter 12)
 
 -------------------------------------------------------------
 - 2017-12-27                                                -
